remove finer in optimize rule
